api/src/test/java/com/example/api/service/ApplyServiceTest.java
package com.example.api.service;

import com.example.api.repository.CouponRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ActiveProfiles;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@ActiveProfiles("docker")
@SpringBootTest
class ApplyServiceTest {

    @Autowired
    private ApplyService applyService;

    @Autowired
    private CouponRepository couponRepository;

    @Test
    public void 한번만응모() {
        applyService.apply(1L); // 쿠폰 발급

        long count = couponRepository.count(); // 쿠폰 카운트

        assertThat(count).isEqualTo(1); // JUnit 테스트 // 한 개의 쿠폰 발급 검증
    }

}